Fixed a bug where the incorrect array indices were being modified in the histogramming process (current_values). Also, t3 modes should be modernized and ready to go.
The next step should be to write some good tests, automating that
process to check for weirdness.

After that, the channels (offsets and whatnot) should be completed.

+++ src/correlate_t2.h
@@ -6,6 +6,7 @@
 #include "combinations.h"
 #include "t2.h"
 #include "options.h"
+#include "types.h"
 
 typedef struct {
 	int order;
@@ -20,7 +21,7 @@ int under_max_distance_t2(t2_t *left, t2_t *right, options_t *options);
 int over_min_distance_t2(t2_t *left, t2_t *right, options_t *options);
 
 int correlate_t2(FILE *in_stream, FILE *out_stream, options_t *options);
-int correlate_t2_block(FILE *out_stream, long long *record_number,
+int correlate_t2_block(FILE *out_stream, int64_t *record_number,
 		t2_queue_t *queue,
 		permutations_t *permutations,
 		offsets_t *offsets, t2_t *correlation_block,
